Update README.md
This code asks the user to input numbers one by one to create a list. The program records the minimum, maximum, and the average of the data set. The second part of this program lets the user enter the a,b, and c variable of quadratic formula with any given number, and will give the user both values of x (including imaginary numbers).